Keep recipe duration authority wasm safe
