+++ crates/oasis7/src/runtime/mod.rs
@@ -125,7 +125,6 @@ pub use gameplay_state::{
     GovernanceVoteState, GovernanceVoteWeightSnapshotState, MetaProgressState,
     WarParticipantOutcome, WarState,
 };
-pub(crate) use modules::canonical_recipe_base_duration_ticks;
 pub use modules::{
     EconomyModuleKind, FactoryBuildDecision, FactoryBuildRequest, FactoryModuleApi,
     FactoryModuleSpec, FactoryProfileV1, GameplayContract, GameplayModuleKind,
