+++ crates/oasis7/src/runtime/modules.rs
@@ -11,37 +11,3 @@ pub use oasis7_wasm_abi::{
     ProductValidationDecision, ProductValidationRequest, RecipeExecutionPlan,
     RecipeExecutionRequest, RecipeModuleApi, RecipeModuleSpec, RecipeProfileV1,
 };
-
-/// Shared base duration authority for the built-in recipe catalog.
-///
-/// Batch count scales material/resource quantities, while each accepted batch
-/// is processed by the same one-tick recipe execution plan.  Keep this lookup
-/// shared by quote construction and runtime action construction so a quote
-/// cannot drift from the `RecipeStarted` duration.
-pub(crate) fn canonical_recipe_base_duration_ticks(recipe_id: &str) -> Option<u32> {
-    match recipe_id.trim().to_ascii_lowercase().as_str() {
-        "recipe.smelter.iron_ingot"
-        | "recipe.iron_ingot"
-        | "recipe.smelter.copper_wire"
-        | "recipe.copper_wire"
-        | "recipe.smelter.polymer_resin"
-        | "recipe.polymer_resin"
-        | "recipe.smelter.alloy_plate"
-        | "recipe.alloy_plate"
-        | "recipe.assembler.gear"
-        | "recipe.gear"
-        | "recipe.assembler.control_chip"
-        | "recipe.control_chip"
-        | "recipe.assembler.motor_mk1"
-        | "recipe.motor_mk1"
-        | "recipe.assembler.logistics_drone"
-        | "recipe.logistics_drone"
-        | "recipe.assembler.sensor_pack"
-        | "recipe.sensor_pack"
-        | "recipe.assembler.module_rack"
-        | "recipe.module_rack"
-        | "recipe.assembler.factory_core"
-        | "recipe.factory_core" => Some(1),
-        _ => None,
-    }
-}
